Added current line in debugger, kinda works...

+++ src/m68k/decode.h
@@ -55,7 +55,7 @@ typedef struct {
     InstructionSize size;
     EffectiveAddress src;
     EffectiveAddress dst;
-    int16_t displacement;
+    int32_t displacement;
 
 } DecodedInstruction;
